requirements.txt
Django==4.2
djangorestframework==3.14.0
django-filter==23.1
drf-yasg==1.21.5
django-environ==0.10.0
django-phonenumber-field==7.0.2
Pillow==9.5.0
